Retrieve SQL Data and Post-process

The workflow for this function is to retrieve the dataset from a source, post-process it according to the value of argument post.op, and, when tgt.name is not NULL, assign it to an object named according to the value of tgt.name

Usage

do.get_data(
  this.conn,
  src.name = NULL,
  tgt.name = NULL,
  this.data = NULL,
  post.op = eval,
  verbose = FALSE,
  promise = FALSE,
  persist.conn = TRUE,
  ...
)

Arguments

this.conn

A DBI-package connection object

src.name

(string) Data source name

tgt.name

(string | NULL) When not NULL, the name of the R workspace object to use when assigning the output. Use format obj_name@env_name to specify a target environment to assign the object

this.data

(expression) When not NULL, an R object for post-operation only

post.op

(object) A function or list of functions to process the retrieved dataset (argument #1) before assigning to workspace

verbose

(logical | FALSE) When TRUE, interim output is printed to console

promise

(logical | FALSE) When TRUE, assignment is done via a call to futureAssign

persist.conn

(logical | TRUE) When FALSE, the connection will be closed at the end of execution

...

Arguments to be used by function do.make_query

Value

Invisibly, the retrieved, post-processed dataset

See also