Export Tabular Data From R to DBMS

export_data serves as a wrapper for dbWriteTable and dbAppendTable. When using append=FALSE, the operation is destructive as the target table is dropped if it exists, so use this function with caution.

export_data(
  this.conn,
  out.data,
  tbl = NULL,
  sch = "dbo",
  append = FALSE,
  persist.conn = TRUE,
  ...
)

Arguments

this.conn

A DBI-package connection object

out.data

The data to write to the database, preferably a data.frame or coercible

tbl

(string) The name of the table to drop (if it exists) and create

sch

(string) The name of the schema in which to create the table

append

(logical) Should the data be appended to existing data rather than replacing existing data?

persist.conn

(logical | TRUE) When FALSE, the connection will be closed at the end of execution

...

Additional arguments passed to dbWriteTable

See also