Make a Snippet

On a new line, make_snippet creates "tag"-like comments referenced by read.snippet (e.g., <snippet: label>...</snippet>). Because of the parsing used, it is important that statements end with a semi-colon (;) as is the case with many other programming languages. The opening "tag" is created as a code section.

make_snippet(..., include.read = TRUE, use.clipboard = FALSE)

Arguments

...

(dots_list) Symbols or words serving as keywords that, taken together, distinguish the snippet from others in the same source document

include.read

(logical) When TRUE, an associated read.snippet command is provided with argument eval = parse

use.clipboard

(logical) When TRUE, the snippet contents are saved to the clipboard and `.Last.value`. This is forcibly set to FALSE for non-Windows operating systems.

Value

When use.clipboard is FALSE, a pair of "tag"-like comments between which code is to be supplied; otherwise, the contents are saved to the clipboard (Windows OS only)

See also