Manual Export of Workspace Objects

The default value for i exports the entire workspace. Unless file is NULL, when i is a vector of names or delimited string of names, the file name becomes 'multiObs'; otherwise, the file name is set to the value of i. When i contains 'all' or '*', regardless of the full content of i, the entire workspace is exported.

save_image(
  ...,
  safe = TRUE,
  env = .GlobalEnv,
  save.dir = getwd(),
  file.name = "",
  use.prefix = TRUE,
  use.timestamp = TRUE,
  prepare = NULL
)

Arguments

...

(dots_list) Names of objects to save given as strings or symbols. Strings may be delimited (c(',', ';', '|', ' '))

safe

(logical | TRUE) Should the pending action be confirmed at the prompt?

env

The environment to search for items

save.dir

(string | getwd()) The directory to save to (not the file name). Use TRUE to interactively choose a save directory.

file.name

(string | "") The name of the file to save, or, when NULL, the value of i if atomic or a predefined name when i is a vector

use.prefix

(logical | TRUE) When TRUE (the default), the file name is prefixed with the value of env

use.timestamp

(logical | TRUE) When TRUE (the default), the file name is appended with a formatted value of Sys.time()

prepare

(language | NULL) A quoted expression that executes before the save action takes place.

Value

A .rdata file, the filename of which being suffixed with a timestamp formatted as "yyyy.mm.dd.hhmmss"

See also