Workflow Manager

workflow_manager is an R6 class that helps to manage workflow-related tasks. The general idea is to set up sets of ordered actions and then selectively execute on-demand.

Public fields

workflows

Holds the workflow sets (expression lists)

log

Holds the history of executed workflow steps Initialize the Workflow Manager

Active bindings

available

Show Available workflow Sets
$available is an R6Class active binding that prints the objects in $workflows, indicating which one is the current workflow (use $help(get) for related information)

current

Show or set the current workflow

Methods

Public methods

workflow_manager$new()

workflow_manager$manage()

workflow_manager$get()

workflow_manager$exec()

workflow_manager$reset()

workflow_manager$clone()

Method new()

$new initializes a new object

Usage

workflow_manager$new(...)

Arguments

...

Not used

Returns

Invisibly, the class environment Manage Workflow Sets

Method manage()

$manage_sets will add or remove workflow sets from $workflows A set is an ordered, named expression list which can be manually defined or created with make.workflow()

Each of the above items can be defined according to taste, but the names must be passed as show above when action is one of 'add', 'update', 'remove', 'delete'. In addition, each element of ... must be supplied as a list (e.g., fitted = list(<objects>))

Usage

workflow_manager$manage(
  action,
  wf_name = private$curnt,
  ...,
  confirm.rm = quote(askYesNo("Remove %s?" %>% sprintf(wf_name))),
  chatty = FALSE
)

Arguments

action

One of add, update, remove, delete, or copy given as strings or symbols:

add

Populate objects into a new object set stored in $workflows

update

Similar to add but to an existing stored object set

remove, delete

Remove a stored object set by name

copy, clone

Copy a stored object set under a new name

wf_name

One or more names to use for saved workflow sets

...

For each term expected in a set, the object data provided in a named list, each element being the same length as wf_name

confirm.rm

(logical) Set to FALSE if you're feeling lucky -- punk!

chatty

(logical \ FALSE) Verbosity flag

Returns

Invisibly, the class environment Get a Saved Set of Workflow Objects

Method get()

$get will retrieve the object in $workflows named in wf_name out and set $current to the environment with the name of that in wf_name

Usage

workflow_manager$get(wf_name)

Arguments

wf_name

A string or symbol of the name of the set to get

Returns

Invisibly, the class environment Execute a Stored Workflow Step

Method exec()

execute.workflow executes quoted expressions referencing read.snippet calls.

Usage

workflow_manager$exec(wf, wf_step, list.only = FALSE, passive = TRUE)

Arguments

wf

(list) The name of the workflow object containing the workflow step expressions

wf_step

(string/symbol) The names of the steps to execute. Tip: label the steps in the order they should execute.

list.only

(logical | FALSE) When TRUE, available workflows and workflow steps are printed to console before the function exits.

passive

(logical | TRUE) When FALSE, workflow steps are interactively selected

Returns

When list.only == TRUE, the listing of workflows and corresponding steps invisibly; otherwise, nothing. Reset the Workflow Manager

Method reset()

$reset will clear out '$workflows' and set '$current' to NULL

Usage

workflow_manager$reset(confirm = TRUE)

Arguments

confirm

(logical | TRUE) Set to FALSE if you're feeling lucky -- punk!

Returns

Invisibly, the class environment

Method clone()

The objects of this class are cloneable with this method.

Usage

workflow_manager$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.